// packages/editor-core/src/templates.generated.ts
export const templateData = {
  "default": {
    "id": "custom",
    "name": "Custom",
    "description": "Custom style derived from user-supplied guidelines.",
    "body_font": "Times New Roman",
    "heading_font": "Times New Roman",
    "body_size_pt": 12,
    "heading_sizes_pt": {
      "1": 14,
      "2": 12,
      "3": 12
    },
    "line_spacing": 2.0,
    "margins_in": {
      "top": 1.0,
      "bottom": 1.0,
      "left": 1.0,
      "right": 1.0
    },
    "columns": 1,
    "heading_case": "title",
    "citation_style": "numeric",
    "required_sections": [
      "abstract",
      "introduction",
      "references"
    ],
    "page_width_in": 8.5,
    "page_height_in": 11.0
  },
  "classes": {
    "ieee-conference": {
      "id": "ieee-conference",
      "name": "IEEE Conference Paper",
      "description": "Two-column IEEE conference paper format (Times New Roman, numeric citations).",
      "body_font": "Times New Roman",
      "heading_font": "Times New Roman",
      "body_size_pt": 10,
      "heading_sizes_pt": {
        "1": 14,
        "2": 11,
        "3": 10
      },
      "line_spacing": 1.0,
      "margins_in": {
        "top": 0.75,
        "bottom": 1.0,
        "left": 0.625,
        "right": 0.625
      },
      "columns": 2,
      "heading_case": "upper",
      "citation_style": "numeric",
      "required_sections": [
        "abstract",
        "introduction",
        "references"
      ],
      "page_width_in": 8.5,
      "page_height_in": 11.0
    },
    "ieee-journal": {
      "id": "ieee-journal",
      "name": "IEEE Transactions Article",
      "description": "Two-column IEEE conference paper format (Times New Roman, numeric citations).",
      "body_font": "Times New Roman",
      "heading_font": "Times New Roman",
      "body_size_pt": 10,
      "heading_sizes_pt": {
        "1": 14,
        "2": 11,
        "3": 10
      },
      "line_spacing": 1.0,
      "margins_in": {
        "top": 0.75,
        "bottom": 1.0,
        "left": 0.625,
        "right": 0.625
      },
      "columns": 2,
      "heading_case": "upper",
      "citation_style": "numeric",
      "required_sections": [
        "abstract",
        "introduction",
        "references"
      ],
      "page_width_in": 8.5,
      "page_height_in": 11.0
    },
    "acm-conference": {
      "id": "acm-conference",
      "name": "ACM Conference Paper",
      "description": "Two-column ACM conference proceedings format (Libertine/Times, numeric citations).",
      "body_font": "Times New Roman",
      "heading_font": "Times New Roman",
      "body_size_pt": 9,
      "heading_sizes_pt": {
        "1": 14,
        "2": 11,
        "3": 10
      },
      "line_spacing": 1.0,
      "margins_in": {
        "top": 0.75,
        "bottom": 1.0,
        "left": 0.75,
        "right": 0.75
      },
      "columns": 2,
      "heading_case": "sentence",
      "citation_style": "numeric",
      "required_sections": [
        "abstract",
        "references"
      ],
      "page_width_in": 8.5,
      "page_height_in": 11.0
    },
    "apa-journal": {
      "id": "apa-journal",
      "name": "APA Journal Manuscript",
      "description": "Single-column APA manuscript format (Times New Roman 12pt, double-spaced, author-date citations).",
      "body_font": "Times New Roman",
      "heading_font": "Times New Roman",
      "body_size_pt": 12,
      "heading_sizes_pt": {
        "1": 12,
        "2": 12,
        "3": 12
      },
      "line_spacing": 2.0,
      "margins_in": {
        "top": 1.0,
        "bottom": 1.0,
        "left": 1.0,
        "right": 1.0
      },
      "columns": 1,
      "heading_case": "title",
      "citation_style": "author-date",
      "required_sections": [
        "abstract",
        "references"
      ],
      "page_width_in": 8.5,
      "page_height_in": 11.0
    },
    "apa-dissertation": {
      "id": "apa-dissertation",
      "name": "APA Dissertation",
      "description": "Single-column APA manuscript format (Times New Roman 12pt, double-spaced, author-date citations).",
      "body_font": "Times New Roman",
      "heading_font": "Times New Roman",
      "body_size_pt": 12,
      "heading_sizes_pt": {
        "1": 12,
        "2": 12,
        "3": 12
      },
      "line_spacing": 2.0,
      "margins_in": {
        "top": 1.0,
        "bottom": 1.0,
        "left": 1.0,
        "right": 1.0
      },
      "columns": 1,
      "heading_case": "title",
      "citation_style": "author-date",
      "required_sections": [
        "abstract",
        "references"
      ],
      "page_width_in": 8.5,
      "page_height_in": 11.0
    }
  }
};